backend/requirements.txt
annotated-types==0.7.0
asgiref==3.8.1
beautifulsoup4==4.13.3
cachetools==5.5.2
certifi==2025.1.31
charset-normalizer==3.4.1
dj-database-url==2.3.0
Django==5.1.6
django-allauth==65.4.1
django-cors-headers==4.7.0
django-environ==0.12.0
djangorestframework==3.15.2
djangorestframework_simplejwt==5.5.0
google==3.0.0
google-ai-generativelanguage==0.6.15
google-api-core==2.24.1
google-api-python-client==2.162.0
google-auth==2.38.0
google-auth-httplib2==0.2.0
google-generativeai==0.8.4
googleapis-common-protos==1.68.0
grpcio==1.70.0
grpcio-status==1.70.0
httplib2==0.22.0
idna==3.10
pillow==11.1.0
proto-plus==1.26.0
protobuf==5.29.3
psycopg==3.2.4
psycopg2-binary==2.9.10
pyasn1==0.6.1
pyasn1_modules==0.4.1
pydantic==2.10.6
pydantic_core==2.27.2
PyJWT==2.9.0
pyparsing==3.2.1
python-dotenv==1.0.1
requests==2.32.3
rsa==4.9
soupsieve==2.6
sqlparse==0.5.3
tqdm==4.67.1
typing_extensions==4.12.2
uritemplate==4.1.1
urllib3==2.3.0
